backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.4.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
Django==5.0
django-anymail==14.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
fonttools==4.61.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
pillow==12.1.0
propcache==0.4.1
psycopg2-binary==2.9.9
pycparser==2.23
pydyf==0.12.1
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
reportlab==4.0.7
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
twilio==9.9.1
typing_extensions==4.15.0
urllib3==2.6.3
weasyprint==67.0
webencodings==0.5.1
yarl==1.22.0
zopfli==0.4.0
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
